+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opModes/camera/openCV/AprilTags/AprilTag.java
@@ -21,6 +21,7 @@
 
 package org.firstinspires.ftc.teamcode.opModes.camera.openCV.AprilTags;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
@@ -33,7 +34,7 @@ import org.openftc.easyopencv.OpenCvCameraRotation;
 import java.util.ArrayList;
 
 @TeleOp
-//@Disabled
+@Disabled
 public class AprilTag extends LinearOpMode {
     OpenCvCamera camera;
     AprilTagDetectionPipeline aprilTagDetectionPipeline;